* Moved necessary base spider attributes to a mixin. * Introduced ArachnadoCrawlSpider to support CrawlSpider based spiders.
